Website: demo fixes (#1069)

+++ examples/core/texture-3d/app.js
@@ -158,7 +158,9 @@ export default class AppAnimationLoop extends AnimationLoop {
   }
 
   onFinalize({gl, cloud}) {
-    cloud.delete();
+    if (cloud) {
+      cloud.delete();
+    }
   }
 
   isSupported() {
